+++ image_process/resize.py
@@ -4,20 +4,20 @@ from ._backend_determine import *
 
 @backend_support('resize', 1)
 def _resize_opencv(img: np.ndarray, width: int, height: int) -> np.ndarray:
-    import cv2
+    import cv2  # type: ignore
     # noinspection PyUnresolvedReferences
     return cv2.resize(img, (width, height), interpolation=cv2.INTER_CUBIC)
 
 
 @backend_support('resize', 0)
 def _resize_skimage(img: np.ndarray, width: int, height: int) -> np.ndarray:
-    import skimage.transform
+    import skimage.transform  # type: ignore
     return np.round(skimage.transform.resize(img, (height, width)) * 255).astype('uint8')
 
 
 @backend_support('resize', 2)
 def _resize_pil(img: np.ndarray, width: int, height: int) -> np.ndarray:
-    from PIL import Image
+    from PIL import Image  # type: ignore
     img_obj = Image.fromarray(img)
     return np.asarray(img_obj.resize((width, height), Image.LANCZOS), dtype='uint8')
 
@@ -45,7 +45,7 @@ def benchmark():
     print('resize pil (png with alpha) time: %f' % t2)
     print('resize skimage (png with alpha) time: %f' % t3)
     # noinspection PyUnboundLocalVariable
-    print('mean abs diff (opencv, pil): %f' % np.mean(np.abs(b.astype(np.float) - c)))
+    print('mean abs diff (opencv, pil): %f' % np.mean(np.abs(b.astype(np.float32) - c)))
     # noinspection PyUnboundLocalVariable
-    print('mean abs diff (pil, skimage): %f' % np.mean(np.abs(c.astype(np.float) - d)))
-    print('mean abs diff (opencv, skimage): %f' % np.mean(np.abs(b.astype(np.float) - d)))
+    print('mean abs diff (pil, skimage): %f' % np.mean(np.abs(c.astype(np.float32) - d)))
+    print('mean abs diff (opencv, skimage): %f' % np.mean(np.abs(b.astype(np.float32) - d)))
